Moved partition data structure
